UCS Satellite Ontology
An ontology of the Union of Concerned Scientists Satellite Database (UCSSO). UCSSO can be used as a local ontology by the UCS. UCSSO is intended as a case-study for demonstrating the development and application of ontology to space object databses for the space situational awareness (SSA) community. It is more specifically points to a domain ontology for SSA and space objects (the SSAO by Rovetto).